+++ tests/testthat/test-epinowcast.R
@@ -28,6 +28,37 @@ test_that("epinowcast preprocesses data and model modules as expected", {
   expect_equal(nowcast[, c("init", "data") := NULL], pobs)
 })
 
+test_that("epinowcast runs using default arguments only", {
+  skip_on_cran()
+  skip_on_local()
+  obs <- germany_covid19_hosp[age_group %in% "00+"][location %in% "DE"] |>
+    enw_filter_report_dates(remove_days = 10) |>
+    enw_filter_reference_dates(include_days = 10)
+  pobs <- enw_preprocess_data(obs, max_delay = 5)
+  nowcast <- suppressMessages(epinowcast(pobs))
+  expect_equal(
+    setdiff(colnames(nowcast), colnames(pobs)),
+    c(
+      "fit", "data", "fit_args", "samples", "max_rhat",
+      "divergent_transitions", "per_divergent_transitions", "max_treedepth",
+      "no_at_max_treedepth", "per_at_max_treedepth", "run_time"
+    )
+  )
+  expect_equal(class(nowcast$fit[[1]])[1], "CmdStanMCMC")
+  expect_type(nowcast$fit_args[[1]], "list")
+  expect_type(nowcast$data[[1]], "list")
+  expect_lt(nowcast$per_divergent_transitions, 0.05)
+  expect_lt(nowcast$max_treedepth, 10)
+  expect_lt(nowcast$max_rhat, 1.05)
+  expect_error(
+    nowcast$fit[[1]]$summary(
+      c("refp_mean_int", "refp_sd_int")
+    ), NA
+  )
+  expect_error(nowcast$fit[[1]]$summary("refp_beta"))
+  expect_error(nowcast$fit[[1]]$summary("rep_beta"))
+})
+
 test_that("epinowcast can fit a simple reporting model", {
   skip_on_cran()
   skip_on_local()